JarEditor: move utils package below internal package
To have a self-contained 'jareditor' namespace.
