+++ bndtools.jareditor/bnd.bnd
@@ -26,7 +26,7 @@ Include-Resource: resources
 	org.eclipse.text
 
 Private-Package: bndtools.jareditor.internal,\
-	bndtools.utils
+	bndtools.jareditor.internal.utils
 	
 Conditional-Package: \
 	aQute.libg.*,\
