+++ bndtools.jareditor/src/bndtools/jareditor/internal/utils/CollectionUtil.java
@@ -1,4 +1,4 @@
-package bndtools.utils;
+package bndtools.jareditor.internal.utils;
 
 import java.util.Collection;
 import java.util.Map;
